#286: Explicitly allow dataclasses_json.cfg.config() argument letter_case to accept LetterCase members
