+++ dataclasses_json/__init__.py
@@ -1,8 +1,8 @@
 # flake8: noqa
 from dataclasses_json.api import (DataClassJsonMixin,
-                                  LetterCase,
                                   dataclass_json)
-from dataclasses_json.cfg import config, global_config, Exclude
+from dataclasses_json.cfg import (config, global_config,
+                                  Exclude, LetterCase)
 from dataclasses_json.undefined import CatchAll, Undefined
 
 __all__ = ['DataClassJsonMixin', 'LetterCase', 'dataclass_json',
